fix: 네비게이션바 내 프로필 버튼 링크 수정
